requirements.txt
pyserial
Pillow
numpy
pystray
